Update rule.h
Changed src and dest port to 16 bit, not sure if they need to be even larger but pretty sure they're represented in values larger than 255.

+++ rule.h
@@ -6,7 +6,7 @@ enum mfw_rule_direction { IN, OUT };
 struct mfw_rule {
     enum mfw_rule_action action;
     enum mfw_rule_direction direction;
-    int8_t protocol;
-    uint8_t srcport;
+    int16_t protocol;
+    int16_t srcport;
     uint8_t destport;
 };
